+++ source/core/actions/example.js
@@ -1,0 +1,7 @@
+export const ADD = "example/ADD"
+export const SUBTRACT = "example/SUBTRACT"
+export const RESET = "example/RESET"
+
+export const addExample = () => ({ type: ADD })
+export const subtractExample = () => ({ type: SUBTRACT })
+export const resetExample = () => ({ type: RESET })
