+++ source/core/reducers/index.js
